components refactory: reusable singleItem and adjustment of props handling
