+++ Views/src/themes/Vibrant/DisplayElements/DisplayElements.css
@@ -15,7 +15,7 @@
 
 .sngl-item{
   width: 100%;
-  border-radius: 1px;
+  border-radius: px;
   height: 100%;
   padding: 0.5rem;
   display: flex;
