+++ Views/src/themes/Vibrant/DisplayElements/DisplayElements.jsx
@@ -1,8 +1,9 @@
 import { useState } from "react";
 import { useFetchAll } from "@hooks/productsHook";
 import { useLongPress } from "@hooks/useLongPress";
-import SingleItem from "../CashierScreen/SingleItem";
+import SingleItem from "../Components/SingleItem";
 import './DisplayElements.css';
+import SlideButton from "./SlideButton";
 
 
 function DisplayElements(){
@@ -19,12 +20,18 @@ function DisplayElements(){
 
     return(
         <div className="display-elements-wrapper">
-            <SingleItem product={label} showButtons={false} />
+            <SingleItem product={label}
+             showButtons={false}
+             ShowPlaceHolder={true}
+             PlaceHolder={active ? "Elimina" : "Modifica"} />
             <div {...longPress} className={!active ? "display-element" : "display-element-DelMode"}>
                 <ul>
                     {products.map((product) => {
                         return(
-                            <SingleItem key={product.id} product={product} showButtons={false}/>
+                            <SingleItem key={product.id}
+                            product={product} 
+                            ShowButtons={true}
+                            Buttons={SlideButton}/>
                         )
                     })}
                 </ul>
