+++ Views/src/themes/Vibrant/DisplayElements/SlideButton.jsx
@@ -1,0 +1,8 @@
+
+export default function SlideButton(){
+    return(
+        <div>
+            Bottone
+        </div>
+    )
+}
